Composer no longer accepts uppercase names.
The package should be re-submitted to Packagist
with a name that follows the naming convention. When
the package was first submitted, there wasn't validation
present at Packagist. Nonetheless, the name doesn't
follow the convention.

Packagist must follow format of either:
* lower-cased/with-dashes
* lower_cased/with_underscores

After changing the name the package can be resent to
Packagist.